requirements.txt
ollama
python-dotenv
firecrawl-py
